chore: refactor vosk into separate helper files
